renamed .cpp source to .c + tabs suck, rewrote with spaced indents
